verifying file exists before renaming

+++ pipeline/luigi/pipeline_common.py
@@ -96,9 +96,10 @@ class DefaultPipelineTask(luigi.Task):
         # renaming files by prefixing filename with "FAILURE_". This way, on rerunning the pipeline the failed task is
         # automatically run again, but instead of just deleting the file, it can still be inspected.
         def _rename_file(path):
-            ts = datetime.datetime.now().strftime("%Y%m%d_%H%M%S")
-            failed_file_name = f"FAILED_{ts}_{os.path.basename(path)}"
-            os.rename(path, os.path.join(os.path.dirname(path), failed_file_name))
+            if os.path.exists(path):
+                ts = datetime.datetime.now().strftime("%Y%m%d_%H%M%S")
+                failed_file_name = f"FAILED_{ts}_{os.path.basename(path)}"
+                os.rename(path, os.path.join(os.path.dirname(path), failed_file_name))
 
         if isinstance(self.output(), luigi.LocalTarget):
             _rename_file(self.output().path)
